move login/signups forms to common
