+++ client/src/states/LoggedIn.jsx
@@ -5,7 +5,7 @@ import { useAppState } from "./AppState";
 import UserDatePosts from "./../components/Post/Posts";
 import Feed from "./../components/Feed/Feed";
 import TopBar from "./../components/TopBar";
-import User from "./../components/User";
+import User from "../components/User/User";
 import NotFound from "../components/common/NotFound";
 import Dashboard from "../components/Dashboard/Dashboard";
